1.验证码业务增加异常信息为"验证码已过期"提示

+++ ruoyi-gateway/src/main/java/com/ruoyi/gateway/service/impl/ValidateCodeServiceImpl.java
@@ -107,8 +107,15 @@ public class ValidateCodeServiceImpl implements ValidateCodeService
         {
             throw new CaptchaException("验证码已失效");
         }
+
         String verifyKey = CacheConstants.CAPTCHA_CODE_KEY + uuid;
         String captcha = redisService.getCacheObject(verifyKey);
+
+        if (StringUtils.isEmpty(captcha))
+        {
+            throw new CaptchaException("验证码已过期");
+        }
+
         redisService.deleteObject(verifyKey);
 
         if (!code.equalsIgnoreCase(captcha))
